[add] add memory pool

+++ src/util/Config.hh
@@ -15,7 +15,7 @@ public:
     ClusterSettings settings;
     string approach;
     bool enable_HDFS;
-    unsigned int block_size; // block size in Bytes
+    uint64_t block_size; // block size in Bytes
     unsigned int num_cmd_handler_thread;
     unsigned int num_cmd_dist_thread;
